Added timeout measure (10h) and minor fixes

+++ benchmark/main.py
@@ -12,36 +12,44 @@ from typing import List
 from benchmark.datasets import DATASETS, get_dataset
 from benchmark.definitions import instantiate_algorithm, get_definitions, list_algorithms, Definition
 from benchmark.algorithms.base.module import BaseClustering
-from benchmark.results import store_results
+from benchmark.results import store_results, test_result_exists
 
 def run_experiment(X: np.array, algo: BaseClustering):
     start = time.time()
+    start_proc = time.process_time()
     algo.cluster(X)
     end = time.time()
-    return end - start, algo.retrieve_dendrogram()
+    end_proc = time.process_time()
+    return end - start - algo.get_overhead_time(), end_proc - start_proc - algo.get_overhead_time(), algo.retrieve_dendrogram(), algo.retrieve_milestones()
 
 
-def run_worker(dataset: str, queue: multiprocessing.Queue) -> None:
+# def run_worker(dataset: str, queue: multiprocessing.Queue) -> None:
+def run_worker(dataset: str, definition: Definition, n_run: int, overwrite: bool = True) -> None:
+    runner = instantiate_algorithm(definition)
+
+    # Immediately return iff file exists and the overwrite flag is false
+    if not overwrite and test_result_exists(dataset, definition.algorithm, repr(runner)): return
+
     X = get_dataset(dataset)
     X = np.array(X["data"])
-    while not queue.empty():
-        definition = queue.get()
+    # while not queue.empty():
+        # definition = queue.get()
+
+    time, time_proc, dendrogram, milestones = run_experiment(X, runner)
+    attrs = {
+        "time": time,
+        "time_proc": time_proc,
+        "ds": dataset,
+        "run": n_run,
+        "algo": definition.algorithm,
+        "params": str(runner)
+    }
+    attrs.update(runner.get_additional())
+    store_results(dataset, definition.algorithm, n_run, 
+                    repr(runner), attrs, dendrogram, milestones)
 
 
-        runner = instantiate_algorithm(definition)
-
-        time, dendrogram = run_experiment(X, runner)
-        attrs = {
-            "time": time,
-            "ds": dataset,
-            "algo": definition.algorithm,
-            "params": str(runner)
-        }
-        attrs.update(runner.get_additional())
-        store_results(dataset, definition.algorithm, 
-                      repr(runner), attrs, dendrogram)
-
-def create_workers_and_execute(dataset: str, definitions: List[Definition]) -> None:
+def create_workers_and_execute(dataset: str, definitions: List[Definition], n_run=1, n_procs=1, overwrite=True) -> None:
     """
     Manages the creation, execution, and termination of worker processes based on provided arguments.
 
@@ -61,18 +69,59 @@ def create_workers_and_execute(dataset: str, definitions: List[Definition]) -> N
     #     raise Exception(
     #         f"Batch mode uses all available CPU resources, --parallelism should be set to 1. (Was: {args.parallelism})"
     #     )
+    
+    cpu_count = multiprocessing.cpu_count()
+    if n_procs > cpu_count - 1:
+       raise Exception(f"Parallelism larger than {cpu_count - 1}! (CPU count minus one)")
 
-    task_queue = multiprocessing.Queue()
-    for run in definitions:
-        task_queue.put(run)
+    timeout_hours = 10
+    timeout_seconds = timeout_hours * 3600
 
-    try:
-        workers = [multiprocessing.Process(target=run_worker, args=(dataset, task_queue))]
-        [worker.start() for worker in workers]
-        [worker.join() for worker in workers]
-    finally:
-        print("Terminating %d workers" % len(workers))
-        [worker.terminate() for worker in workers]
+    if n_procs == 1:
+        # task_queue = multiprocessing.Queue()
+        for run in definitions:
+            # task_queue.put(run)
+
+            try:
+                workers = [multiprocessing.Process(target=run_worker, args=(dataset, run), kwargs=dict(n_run=n_run, overwrite=overwrite))]
+                [worker.start() for worker in workers]
+                [worker.join(timeout=timeout_seconds) for worker in workers] # Timeout of 10 hours
+
+                for worker in workers:
+                    if worker.is_alive():
+                        print("Timeout reached. Terminating worker...")
+                        worker.terminate()
+                        worker.join()
+                    else:
+                        print("Worker completed within time.")
+            finally:
+                print("Terminating %d workers" % len(workers))
+                [worker.terminate() for worker in workers]
+    else:
+        # Encapsulate the above code in a "nicer" way for a thread worker
+        def pool_worker(run):
+            try:
+                worker = multiprocessing.Process(target=run_worker, args=(dataset, run), kwargs=dict(n_run=n_run, overwrite=overwrite))
+                worker.start()
+                worker.join(timeout=timeout_seconds) # Timeout of 10 hours
+                if worker.is_alive(): raise TimeoutError()
+                print("Worker completed within time.")
+            except Exception as e:
+                if worker.is_alive():
+                    print("Timeout reached. Terminating worker...")
+                    worker.terminate()
+                    if worker.is_alive():
+                        kill_result = os.system(f"kill {worker.pid}")
+                        if kill_result: print(f"Warning: Failed to kill child process {worker.pid}")
+                    else: worker.join()
+                if type(e) != TimeoutError:
+                    import traceback
+                    traceback.print_exception(e)
+        # Use a thread pool to have a managed number of executions running at any one time
+        pool = multiprocessing.pool.ThreadPool(n_procs)
+        pool.map(pool_worker, definitions)
+        pool.close()
+
 
 def main():
     parser = argparse.ArgumentParser(
@@ -102,6 +151,29 @@ def main():
         help='only prepare the dataset'
     )
 
+    parser.add_argument(
+        '-r',
+        '--run',
+        type=int,
+        default=1
+    )
+
+    parser.add_argument(
+        '-o',
+        '--overwrite',
+        type=bool,
+        default=True,
+        help='whether or not to overwrite or skip existing results'
+    )
+
+    parser.add_argument(
+        '-p',
+        '--procs',
+        type=int,
+        default=1,
+        help="the number of processes to use for experiments",
+    )
+
     args = parser.parse_args()
 
 
@@ -124,4 +196,4 @@ def main():
     if args.prepare:
         exit(0)
 
-    create_workers_and_execute(args.dataset, definitions)
+    create_workers_and_execute(args.dataset, definitions, n_run=args.run, n_procs=args.procs, overwrite=args.overwrite)
